requirements.txt
yfinance>=0.2.31
twelvedata>=1.2.0
finnhub-python>=2.4.0
fredapi>=0.5.0
newsapi-python>=0.2.7
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
numpy>=1.24.0
textblob>=0.17.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
pyyaml>=6.0
tenacity>=8.2.0
pytz>=2023.3

# LLM providers (optional — install whichever you have a key for)
anthropic>=0.39.0
openai>=1.50.0
google-generativeai>=0.8.0

# Web UI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
